Restructured domain management and domain alias commands
