+++ keepercommander/commands/enterprise.py
@@ -17,7 +17,6 @@ import itertools
 import json
 import logging
 import os
-import re
 import string
 import time
 from argparse import RawTextHelpFormatter
@@ -36,6 +35,7 @@ from ..attachment import FileUploadTask
 
 from . import aram, audit_alerts, security_audit
 from . import compliance
+from . import domain_management
 from .aram import ActionReportCommand, API_EVENT_SUMMARY_ROW_LIMIT
 from .base import user_choice, suppress_exit, raise_parse_exception, dump_report_data, Command, field_to_title, \
     report_output_parser
@@ -60,7 +60,7 @@ def register_commands(commands):
     commands['team-approve'] = TeamApproveCommand()
     commands['device-approve'] = DeviceApproveCommand()
     commands['transfer-user'] = EnterpriseTransferUserCommand()
-    commands['domain'] = DomainCommand()
+    domain_management.register_commands(commands)
     commands['audit-log'] = aram.AuditLogCommand()
     commands['audit-report'] = aram.AuditReportCommand()
     commands['aging-report'] = aram.AgingReportCommand()
@@ -84,21 +84,17 @@ def register_command_info(aliases, command_info):
     aliases['et'] = 'enterprise-team'
     aliases['esr'] = 'external-shares-report'
     aliases['tu'] = 'transfer-user'
-    aliases['dl'] = ('domain', 'list')
-    aliases['dr'] = ('domain', 'reserve')
-    aliases['dal'] = ('domain', 'alias', 'list')
-    aliases['dac'] = ('domain', 'alias', 'create')
-    aliases['dad'] = ('domain', 'alias', 'delete')
 
     for p in [enterprise_data_parser, enterprise_info_parser, enterprise_node_parser, enterprise_user_parser,
               enterprise_role_parser, enterprise_team_parser, transfer_user_parser,
               enterprise_push_parser, team_approve_parser, device_approve_parser,
               aram.audit_log_parser, aram.audit_report_parser, aram.aging_report_parser, aram.action_report_parser,
-              user_report_parser, domain_parser]:
+              user_report_parser]:
         command_info[p.prog] = p.description
 
     command_info['audit-alert'] = 'Manage audit alerts and notifications'
 
+    domain_management.register_command_info(aliases, command_info)
     compliance.register_command_info(aliases, command_info)
     security_audit.register_command_info(aliases, command_info)
 
@@ -288,89 +284,6 @@ user_report_parser.add_argument('-l', '--last-login', dest='last_login', action=
                                 help='simplify report to include only last-login-related info')
 user_report_parser.error = raise_parse_exception
 user_report_parser.exit = suppress_exit
-
-domain_parser = argparse.ArgumentParser(prog='domain', description='Manage enterprise domains')
-domain_parser.error = raise_parse_exception
-domain_parser.exit = suppress_exit
-
-domain_subparsers = domain_parser.add_subparsers(dest='subcommand', help='Domain subcommands', metavar='{list,reserve,alias}')
-
-domain_list_parser = domain_subparsers.add_parser('list', parents=[report_output_parser],
-                                                   help='List all reserved domains for the enterprise',
-                                                   description='List all reserved domains for the enterprise.')
-domain_list_parser.error = raise_parse_exception
-domain_list_parser.exit = suppress_exit
-
-domain_reserve_parser = domain_subparsers.add_parser('reserve',
-                                                      formatter_class=argparse.RawTextHelpFormatter,
-                                                      help='Reserve and manage domains',
-                                                      description='Reserve and manage domains for the enterprise.\n\n'
-                                                      'Process:\n'
-                                                      ' 1. Use --action token to get DNS verification token\n'
-                                                      ' 2. Add TXT record to your DNS\n'
-                                                      ' 3. Use --action add to complete reservation\n'
-                                                      ' 4. Use --action delete to remove domain')
-domain_reserve_parser.add_argument('--action', dest='action', required=True,
-                                    choices=['token', 'add', 'delete'],
-                                    help='Action to perform: token (get verification token), add (add domain after verification), delete (remove domain)')
-domain_reserve_parser.add_argument('--domain', dest='domain', required=True,
-                                    help='Domain name to reserve')
-domain_reserve_parser.add_argument('--format', dest='format', action='store', choices=['text', 'json'],
-                                    default='text', help='Output format.')
-domain_reserve_parser.add_argument('--force', dest='force', action='store_true',
-                                    help='Skip confirmation prompt for delete action')
-domain_reserve_parser.error = raise_parse_exception
-domain_reserve_parser.exit = suppress_exit
-
-# --- domain alias <sub-subcommand> ---
-domain_alias_parser = domain_subparsers.add_parser('alias',
-                                                    help='Manage domain aliases for the enterprise',
-                                                    description='Manage domain aliases for the enterprise.')
-domain_alias_parser.error = raise_parse_exception
-domain_alias_parser.exit = suppress_exit
-
-domain_alias_subparsers = domain_alias_parser.add_subparsers(
-    dest='alias_subcommand', help='Alias subcommands', metavar='{list,create,delete}'
-)
-
-domain_alias_list_parser = domain_alias_subparsers.add_parser('list', parents=[report_output_parser],
-                                                               help='List domain aliases for the enterprise',
-                                                               description='List domain aliases for the enterprise.')
-domain_alias_list_parser.error = raise_parse_exception
-domain_alias_list_parser.exit = suppress_exit
-
-domain_alias_create_parser = domain_alias_subparsers.add_parser('create',
-                                                                 formatter_class=argparse.RawTextHelpFormatter,
-                                                                 help='Create domain aliases for the enterprise',
-                                                                 description='Create aliases for domains owned by the enterprise.\n\n'
-                                                                 'The enterprise must own the domain before an alias can be created.\n'
-                                                                 'Requires Admin with "Manage Users" permission.')
-domain_alias_create_parser.add_argument('--domain', dest='domain', required=True,
-                                        help='Domain name to create alias for (must be owned by the enterprise)')
-domain_alias_create_parser.add_argument('--alias', dest='alias', required=True, action='append',
-                                        help='Alias to create for the domain (can be specified multiple times)')
-domain_alias_create_parser.add_argument('--format', dest='format', action='store', choices=['text', 'json'],
-                                        default='text', help='Output format.')
-domain_alias_create_parser.error = raise_parse_exception
-domain_alias_create_parser.exit = suppress_exit
-
-domain_alias_delete_parser = domain_alias_subparsers.add_parser('delete',
-                                                                 formatter_class=argparse.RawTextHelpFormatter,
-                                                                 help='Delete domain aliases for the enterprise',
-                                                                 description='Delete aliases for domains owned by the enterprise.\n\n'
-                                                                 'Only previously created aliases can be deleted.\n'
-                                                                 'Requires Admin with "Manage Users" permission.')
-domain_alias_delete_parser.add_argument('--domain', dest='domain', required=True,
-                                        help='Domain name whose alias to delete')
-domain_alias_delete_parser.add_argument('--alias', dest='alias', required=True, action='append',
-                                        help='Alias to delete (can be specified multiple times)')
-domain_alias_delete_parser.add_argument('--format', dest='format', action='store', choices=['text', 'json'],
-                                        default='text', help='Output format.')
-domain_alias_delete_parser.add_argument('--force', dest='force', action='store_true',
-                                        help='Skip confirmation prompt')
-domain_alias_delete_parser.error = raise_parse_exception
-domain_alias_delete_parser.exit = suppress_exit
-
 
 _DEFAULT_PASSWORD_COMPLEXITY = """[
   {
@@ -4344,698 +4257,3 @@ class DeviceApproveCommand(EnterpriseCommand):
                 ])
             rows.sort(key=lambda x: x[0])
             return dump_report_data(rows, headers, fmt=kwargs.get('format'), filename=kwargs.get('output'))
-
-
-class DomainManagementHelper:
-    
-    # Domain validation constants
-    MAX_DOMAIN_LENGTH = 253
-    MAX_LABEL_LENGTH = 63
-    MIN_TLD_LENGTH = 2
-    
-    NOTICE_MSG = 'Notice: This feature is not in production yet. It will be available soon.'    
-    DOMAIN_PATTERN = r'^(?!-)([a-z0-9-]{1,63})(?<!-)(\.(?!-)([a-z0-9-]{1,63})(?<!-))*$'
-    
-    ERROR_MESSAGES = {
-        'bad_request': 'Domain not specified or invalid',
-        'access_denied': 'Access denied: You must be a Root Admin to manage domains',
-        'forbidden': 'Access denied: You must be a Root Admin to manage domains',
-        'domain_already_taken': 'Domain "{domain}" is already reserved by a different enterprise',
-        'verification_failed': 'DNS verification failed for domain "{domain}". Please ensure the TXT record is correctly added and DNS has propagated (may take up to 48 hours).',
-        'invalid_domain': 'Invalid domain format: "{domain}"',
-        'rate_limit': 'Too many requests. Please wait a moment and try again.',
-        'too_many_requests': 'Too many requests. Please wait a moment and try again.',
-    }
-
-    ALIAS_ACCESS_DENIED_MSG = 'Access denied: You must be an Admin with "Manage Users" permission to manage domain aliases.'
-
-    CREATE_ALIAS_STATUS_MESSAGES = {
-        0: 'Success',
-        1: 'Duplicate; already exists',
-        2: 'Not allowed; domain or alias not owned by the enterprise',
-    }
-
-    DELETE_ALIAS_STATUS_MESSAGES = {
-        0: 'Success',
-        1: 'Not allowed or does not exist',
-    }
-
-    @staticmethod
-    def get_error_code(e):
-        """Return the result_code from a KeeperApiError, or 'Unknown' if absent."""
-        return e.result_code if hasattr(e, 'result_code') else 'Unknown'
-
-    @staticmethod
-    def is_feature_unavailable(error_code):
-        return error_code == 404 or error_code == '404' or error_code == 'invalid_path_or_method'
-    
-    @staticmethod
-    def handle_unavailable_feature(output_format='text'):
-        if output_format == 'json':
-            notice_output = {
-                'notice': DomainManagementHelper.NOTICE_MSG,
-            }
-            return json.dumps(notice_output, indent=2)
-        else:
-            logging.warning(f"{bcolors.WARNING}{DomainManagementHelper.NOTICE_MSG}{bcolors.ENDC}")
-            return None
-    
-    @staticmethod
-    def output_error(error_msg, output_format='text', **additional_fields):
-        if output_format == 'json':
-            error_output = {'error': error_msg}
-            error_output.update(additional_fields)
-            print(json.dumps(error_output, indent=2))
-        else:
-            logging.error(error_msg)
-    
-    @staticmethod
-    def validate_domain(domain):
-        """
-        Validate domain name format and requirements.
-        
-        Args:
-            domain: Domain name to validate
-            
-        Returns:
-            tuple: (is_valid: bool, normalized_domain: str, error_message: str or None)
-        """
-        if not domain:
-            return False, None, 'Domain name is required'
-        
-        domain = domain.strip().lower()
-        
-        if not domain or len(domain) > DomainManagementHelper.MAX_DOMAIN_LENGTH:
-            return False, domain, f'Invalid domain name: must be between 1 and {DomainManagementHelper.MAX_DOMAIN_LENGTH} characters'
-        
-        if not re.match(DomainManagementHelper.DOMAIN_PATTERN, domain):
-            return False, domain, 'Invalid domain format: domain must contain only letters, numbers, hyphens, and dots'
-        
-        if '.' not in domain:
-            return False, domain, 'Invalid domain: must contain at least one dot (e.g., example.com)'
-        
-        # Check each label
-        labels = domain.split('.')
-        for label in labels:
-            if len(label) > DomainManagementHelper.MAX_LABEL_LENGTH:
-                return False, domain, f'Invalid domain: label "{label}" exceeds {DomainManagementHelper.MAX_LABEL_LENGTH} characters'
-        
-        # Check TLD length
-        if len(labels[-1]) < DomainManagementHelper.MIN_TLD_LENGTH:
-            return False, domain, f'Invalid domain: TLD must be at least {DomainManagementHelper.MIN_TLD_LENGTH} characters'
-        
-        return True, domain, None
-    
-    @staticmethod
-    def get_error_message(error_code, domain, action):
-        """Get user-friendly error message using class constant dictionary."""
-        
-        if error_code == 'invalid_token' and action == 'add':
-            return f'Failed to verify domain "{domain}". Please ensure you have added the TXT record with the correct token to your DNS settings and try again.'
-        
-        if error_code in ('exists', 'domain_exists'):
-            if action == 'token':
-                return f'Domain "{domain}" already exists in the enterprise. Use action "delete" to remove it first.'
-            elif action == 'add':
-                return f'Domain "{domain}" already exists in the enterprise. It may have already been added successfully.'
-        
-        if error_code in ('not_exists', 'domain_not_found', 'doesnt_exist') and action == 'delete':
-            return f'Domain "{domain}" does not exist. Use action "token" to start the domain reservation process.'
-        
-        message_template = DomainManagementHelper.ERROR_MESSAGES.get(error_code)
-        
-        if message_template:
-            if '{domain}' in message_template:
-                return message_template.format(domain=domain)
-            return message_template
-        
-        return f'Unable to {action} domain "{domain}". Please try again or contact support if the issue persists.'
-
-    @staticmethod
-    def validate_aliases(domain, aliases, output_format):
-        """
-        Validate a list of alias names and check none equals the domain.
-
-        Returns a list of normalized alias strings on success, or None if any
-        validation fails (error is already reported via output_error).
-        """
-        normalized = []
-        for alias_name in aliases:
-            valid, normalized_alias, err = DomainManagementHelper.validate_domain(alias_name)
-            if not valid:
-                DomainManagementHelper.output_error(f'{err}', output_format)
-                return None
-            if normalized_alias == domain:
-                DomainManagementHelper.output_error(
-                    f'Alias cannot be the same as the domain ("{domain}").', output_format
-                )
-                return None
-            normalized.append(normalized_alias)
-        return normalized
-
-    @staticmethod
-    def build_alias_request(domain, normalized_aliases):
-        """Build a DomainAliasRequest protobuf from a validated domain and alias list."""
-        rq = enterprise_pb2.DomainAliasRequest()
-        for alias in normalized_aliases:
-            da = enterprise_pb2.DomainAlias()
-            da.domain = domain
-            da.alias = alias
-            rq.domainAlias.append(da)
-        return rq
-
-    @staticmethod
-    def render_alias_response(rs, output_format, kwargs, status_messages=None, action=None):
-        """
-        Render a DomainAliasResponse as text or JSON.
-
-        When status_messages is None the response is treated as a listing
-        (domain + alias only).  When provided, a simple message is shown per
-        alias for text format, or a status object for JSON format.
-        action should be 'create' or 'delete' when status_messages is provided.
-        """
-        if status_messages is not None:
-            if output_format == 'json':
-                results = [
-                    {
-                        'domain': da.domain,
-                        'alias': da.alias,
-                        'status': da.status,
-                        'status_message': status_messages.get(da.status, f'Unknown status: {da.status}'),
-                    }
-                    for da in rs.domainAlias
-                ]
-                print(json.dumps(results, indent=2))
-            else:
-                for da in rs.domainAlias:
-                    status_msg = status_messages.get(da.status, f'Unknown ({da.status})')
-                    if da.status == 0:
-                        if action == 'create':
-                            logging.info(f"Created domain alias '{da.alias}' for domain '{da.domain}'")
-                        elif action == 'delete':
-                            logging.info(f"Deleted domain alias '{da.alias}' for domain '{da.domain}'")
-                        else:
-                            logging.info(f"Domain alias '{da.alias}' for domain '{da.domain}': {status_msg}")
-                    else:
-                        logging.error(f"Domain alias '{da.alias}' for domain '{da.domain}': {status_msg}")
-        else:
-            if output_format == 'json':
-                results = [{'domain': da.domain, 'alias': da.alias} for da in rs.domainAlias]
-                print(json.dumps(results, indent=2))
-            else:
-                headers = ['Domain', 'Alias']
-                table = [[da.domain, da.alias] for da in rs.domainAlias]
-                return dump_report_data(table, headers, fmt=output_format, filename=kwargs.get('output'))
-
-    @staticmethod
-    def handle_alias_api_error(e, output_format, operation):
-        """
-        Shared KeeperApiError handler for alias commands.
-
-        Handles feature-unavailable and access_denied cases, logs unexpected
-        errors, and re-raises them so the caller can surface a traceback.
-        """
-        error_code = DomainManagementHelper.get_error_code(e)
-
-        if DomainManagementHelper.is_feature_unavailable(error_code):
-            result = DomainManagementHelper.handle_unavailable_feature(output_format)
-            if result:
-                print(result)
-            return
-
-        if error_code == 'access_denied':
-            DomainManagementHelper.output_error(
-                DomainManagementHelper.ALIAS_ACCESS_DENIED_MSG, output_format
-            )
-            return
-
-        logging.error(f'Error {operation} domain aliases: {e}')
-        raise
-
-    @staticmethod
-    def handle_invalid_subcommand(subcommand, output_format='text'):
-        """
-        Handle invalid subcommand error with user-friendly message.
-        
-        Args:
-            subcommand: The invalid subcommand provided by the user
-            output_format: Output format (text or json)
-        """
-        error_message = (
-            f"Invalid subcommand: '{subcommand}'. "
-            f"Use 'domain --help' for more information."
-        )
-        
-        if output_format == 'json':
-            error_output = {
-                'error': error_message,
-            }
-            print(json.dumps(error_output, indent=2))
-        else:
-            logging.error(error_message)
-
-
-class DomainCommand(EnterpriseCommand):
-    def __init__(self):
-        super().__init__()
-        self.list_cmd = ListDomainsCommand()
-        self.reserve_cmd = ReserveDomainCommand()
-        self.alias_cmd = DomainAliasCommand()
-
-    def get_parser(self):
-        return domain_parser
-
-    def execute_args(self, params, args, **kwargs):
-        import shlex
-        from .base import ParseError, expand_cmd_args, normalize_output_param
-        
-        try:
-            d = {}
-            d.update(kwargs)
-            self.extra_parameters = ''
-            parser = self._get_parser_safe()
-            envvars = params.environment_variables
-            args = '' if args is None else args
-            
-            if parser:
-                args = expand_cmd_args(args, envvars)
-                args = normalize_output_param(args)
-                opts = parser.parse_args(shlex.split(args))
-                d.update(opts.__dict__)
-
-            return self.execute(params, **d)
-            
-        except ParseError as e:
-            error_str = str(e)
-            if 'invalid choice' in error_str:
-                match = re.search(r"invalid choice: '([^']+)'", error_str)
-                if match:
-                    invalid_cmd = match.group(1)
-                    output_format = kwargs.get('format', 'text')
-                    DomainManagementHelper.handle_invalid_subcommand(invalid_cmd, output_format)
-                    return None
-            logging.error(error_str)
-            return None
-
-    def execute(self, params, **kwargs):
-        subcommand = kwargs.get('subcommand')
-
-        if not subcommand:
-            self.get_parser().print_help()
-            return
-
-        if subcommand == 'list':
-            return self.list_cmd.execute(params, **kwargs)
-        elif subcommand == 'reserve':
-            return self.reserve_cmd.execute(params, **kwargs)
-        elif subcommand == 'alias':
-            return self.alias_cmd.execute(params, **kwargs)
-        else:
-            output_format = kwargs.get('format', 'text')
-            DomainManagementHelper.handle_invalid_subcommand(subcommand, output_format)
-            return None
-
-
-class ListDomainsCommand(EnterpriseCommand):
-    def get_parser(self):
-        return domain_list_parser
-
-    def execute(self, params, **kwargs):
-        try:
-            rs = api.communicate_rest(
-                params, 
-                None,  
-                'enterprise/list_domains',
-                rs_type=enterprise_pb2.ListDomainsResponse
-            )
-            
-            fmt = kwargs.get('format', '')    
-            
-            if not rs.domain:
-                logging.info('No reserved domains found for this enterprise.')
-                return
-            
-            if fmt == 'json':
-                domains_list = list(rs.domain)
-                print(json.dumps(domains_list, indent=2))
-            else:
-                headers = ['Domain Name']
-                table = [[domain] for domain in rs.domain]
-                return dump_report_data(table, headers, fmt=fmt, filename=kwargs.get('output'))
-                
-        except KeeperApiError as e:
-            error_code = e.result_code if hasattr(e, 'result_code') else 'Unknown'
-            
-            if DomainManagementHelper.is_feature_unavailable(error_code):
-                result = DomainManagementHelper.handle_unavailable_feature(kwargs.get('format') or 'text')
-                if result:
-                    print(result)
-                return
-            
-            logging.error(f'Error listing domains: {e}')
-            raise
-
-
-class DomainAliasCommand(EnterpriseCommand):
-    """Routes `domain alias <list|create|delete>` to the appropriate command."""
-
-    def __init__(self):
-        super().__init__()
-        self.list_cmd = GetDomainAliasCommand()
-        self.create_cmd = CreateDomainAliasCommand()
-        self.delete_cmd = DeleteDomainAliasCommand()
-
-    def get_parser(self):
-        return domain_alias_parser
-
-    def execute(self, params, **kwargs):
-        alias_subcommand = kwargs.get('alias_subcommand')
-
-        if not alias_subcommand:
-            self.get_parser().print_help()
-            return
-
-        if alias_subcommand == 'list':
-            return self.list_cmd.execute(params, **kwargs)
-        elif alias_subcommand == 'create':
-            return self.create_cmd.execute(params, **kwargs)
-        elif alias_subcommand == 'delete':
-            return self.delete_cmd.execute(params, **kwargs)
-        else:
-            output_format = kwargs.get('format', 'text')
-            DomainManagementHelper.handle_invalid_subcommand(
-                f'alias {alias_subcommand}', output_format
-            )
-            return None
-
-
-class GetDomainAliasCommand(EnterpriseCommand):
-    def get_parser(self):
-        return domain_alias_list_parser
-
-    def execute(self, params, **kwargs):
-        output_format = kwargs.get('format') or 'text'
-        try:
-            rs = api.communicate_rest(
-                params,
-                None,
-                'enterprise/get_domain_alias',
-                rs_type=enterprise_pb2.DomainAliasResponse
-            )
-
-            if not rs.domainAlias:
-                logging.info('No domain aliases found for this enterprise.')
-                return
-
-            return DomainManagementHelper.render_alias_response(rs, output_format, kwargs)
-
-        except KeeperApiError as e:
-            DomainManagementHelper.handle_alias_api_error(e, output_format, 'retrieving')
-
-
-class CreateDomainAliasCommand(EnterpriseCommand):
-
-    def get_parser(self):
-        return domain_alias_create_parser
-
-    def execute(self, params, **kwargs):
-        domain = kwargs.get('domain', '')
-        aliases = kwargs.get('alias', [])
-        output_format = kwargs.get('format', 'text')
-
-        if not domain:
-            DomainManagementHelper.output_error('Domain name is required.', output_format)
-            return
-        if not aliases:
-            DomainManagementHelper.output_error('At least one alias is required.', output_format)
-            return
-
-        is_valid, domain, error_msg = DomainManagementHelper.validate_domain(domain)
-        if not is_valid:
-            DomainManagementHelper.output_error(error_msg, output_format)
-            return
-
-        normalized_aliases = DomainManagementHelper.validate_aliases(domain, aliases, output_format)
-        if normalized_aliases is None:
-            return
-
-        try:
-            rq = DomainManagementHelper.build_alias_request(domain, normalized_aliases)
-            rs = api.communicate_rest(
-                params,
-                rq,
-                'enterprise/create_domain_alias',
-                rs_type=enterprise_pb2.DomainAliasResponse
-            )
-            return DomainManagementHelper.render_alias_response(
-                rs, output_format, kwargs,
-                status_messages=DomainManagementHelper.CREATE_ALIAS_STATUS_MESSAGES,
-                action='create'
-            )
-
-        except KeeperApiError as e:
-            DomainManagementHelper.handle_alias_api_error(e, output_format, 'creating')
-
-
-class DeleteDomainAliasCommand(EnterpriseCommand):
-
-    def get_parser(self):
-        return domain_alias_delete_parser
-
-    def execute(self, params, **kwargs):
-        domain = kwargs.get('domain', '')
-        aliases = kwargs.get('alias', [])
-        output_format = kwargs.get('format', 'text')
-        force = kwargs.get('force', False)
-
-        if not domain:
-            DomainManagementHelper.output_error('Domain name is required.', output_format)
-            return
-        if not aliases:
-            DomainManagementHelper.output_error('At least one alias is required.', output_format)
-            return
-
-        is_valid, domain, error_msg = DomainManagementHelper.validate_domain(domain)
-        if not is_valid:
-            DomainManagementHelper.output_error(error_msg, output_format)
-            return
-
-        normalized_aliases = DomainManagementHelper.validate_aliases(domain, aliases, output_format)
-        if normalized_aliases is None:
-            return
-
-        if not force:
-            alias_list_str = ', '.join(normalized_aliases)
-            try:
-                confirm = input(
-                    f'Are you sure you want to delete alias(es) [{alias_list_str}] for domain "{domain}"? (y/N): '
-                )
-            except (KeyboardInterrupt, EOFError):
-                logging.info('Delete cancelled.')
-                return
-            if confirm.strip().lower() not in ('y', 'yes'):
-                logging.info('Delete cancelled.')
-                return
-
-        try:
-            rq = DomainManagementHelper.build_alias_request(domain, normalized_aliases)
-            rs = api.communicate_rest(
-                params,
-                rq,
-                'enterprise/delete_domain_alias',
-                rs_type=enterprise_pb2.DomainAliasResponse
-            )
-            return DomainManagementHelper.render_alias_response(
-                rs, output_format, kwargs,
-                status_messages=DomainManagementHelper.DELETE_ALIAS_STATUS_MESSAGES,
-                action='delete'
-            )
-
-        except KeeperApiError as e:
-            DomainManagementHelper.handle_alias_api_error(e, output_format, 'deleting')
-
-
-class ReserveDomainCommand(EnterpriseCommand):
-    
-    ACTION_MAP = {
-        'token': enterprise_pb2.DOMAIN_TOKEN,
-        'add': enterprise_pb2.DOMAIN_ADD,
-        'delete': enterprise_pb2.DOMAIN_DELETE
-    }
-    
-    def get_parser(self):
-        return domain_reserve_parser
-
-    def execute(self, params, **kwargs):
-        action = kwargs.get('action')
-        domain = kwargs.get('domain')
-        output_format = kwargs.get('format', 'text')
-        force = kwargs.get('force', False)
-        
-        if not self._validate_inputs(action, domain, output_format):
-            return
-        
-        is_valid, domain, error_msg = DomainManagementHelper.validate_domain(domain)
-        if not is_valid:
-            DomainManagementHelper.output_error(error_msg, output_format, domain=domain or '', status='failed')
-            return
-        
-        try:
-            result = self._execute_action(params, action, domain, output_format, force=force)
-            if result:
-                return result
-                
-        except KeeperApiError as e:
-            return self._handle_api_error(e, domain, action, output_format)
-                
-        except Exception as e:
-            error_msg = f'Unexpected error: {str(e)}'
-            DomainManagementHelper.output_error(error_msg, output_format, domain=domain, action=action)
-            logging.debug(f'Exception details: {e}', exc_info=True)
-    
-    def _validate_inputs(self, action, domain, output_format):
-        """Validate action and domain inputs."""
-        if not action:
-            DomainManagementHelper.output_error('Action is required', output_format, status='failed')
-            return False
-        
-        if action not in self.ACTION_MAP:
-            DomainManagementHelper.output_error(
-                f'Invalid action: {action}. Must be one of: {", ".join(self.ACTION_MAP.keys())}',
-                output_format,
-                status='failed'
-            )
-            return False
-        
-        if not domain:
-            DomainManagementHelper.output_error('Domain is required', output_format, status='failed')
-            return False
-        
-        return True
-    
-    def _execute_action(self, params, action, domain, output_format, force=False):
-        """Execute the specified domain action after validation."""
-        if not action or not domain:
-            DomainManagementHelper.output_error('Action and domain are required', output_format, status='failed')
-            return
-        
-        rq = self._create_request(action, domain)
-        
-        if action == 'token':
-            return self._handle_token_action(params, rq, domain, output_format)
-        elif action == 'add':
-            return self._handle_add_action(params, rq, domain, output_format)
-        elif action == 'delete':
-            return self._handle_delete_action(params, rq, domain, output_format, force=force)
-    
-    def _create_request(self, action, domain):
-        rq = enterprise_pb2.ReserveDomainRequest()
-        rq.reserveDomainAction = self.ACTION_MAP[action]
-        rq.domain = domain       
-        
-        return rq
-    
-    def _handle_token_action(self, params, rq, domain, output_format):
-        rs = api.communicate_rest(
-            params, 
-            rq,
-            'enterprise/reserve_domain',
-            rs_type=enterprise_pb2.ReserveDomainResponse
-        )
-        
-        if not rs or not hasattr(rs, 'token') or not rs.token:
-            DomainManagementHelper.output_error(
-                'Failed to generate token: empty response from server',
-                output_format,
-                domain=domain,
-            )
-            return
-        
-        if output_format == 'json':
-            return json.dumps({'token': rs.token, 'domain': domain}, indent=2)
-        
-        self._display_token_instructions(domain, rs.token)
-    
-    def _handle_add_action(self, params, rq, domain, output_format):
-        api.communicate_rest(params, rq, 'enterprise/reserve_domain')
-        
-        if output_format == 'json':
-            return json.dumps({
-                'message': 'Domain successfully added to enterprise',
-                'domain': domain,
-                'action': 'add',
-            }, indent=2)
-        
-        logging.info(f'Domain "{domain}" has been reserved for the enterprise')
-        self._refresh_enterprise_data(params, 'added')
-    
-    def _handle_delete_action(self, params, rq, domain, output_format, force=False):
-        """Handle domain deletion with optional confirmation."""
-        if not force and output_format != 'json':
-            domain_exists = self._check_domain_exists(params, domain)
-            if not domain_exists:
-                pass
-            else:
-                confirm = input(f'\n{bcolors.WARNING}Are you sure you want to delete domain "{domain}"? (y/n): {bcolors.ENDC}')
-                if confirm.lower() not in ['yes', 'y']:
-                    logging.info('Domain deletion cancelled')
-                    return
-        
-        api.communicate_rest(params, rq, 'enterprise/reserve_domain')
-        
-        if output_format == 'json':
-            return json.dumps({
-                'message': 'Domain removed from enterprise',
-                'domain': domain,
-                'action': 'delete',
-            }, indent=2)
-        
-        logging.info(f'Domain "{domain}" has been removed from the enterprise')
-        self._refresh_enterprise_data(params, 'removed')
-    
-    def _check_domain_exists(self, params, domain):
-        """Check if a domain exists in the enterprise."""
-        rs = api.communicate_rest(
-            params,
-            None,
-            'enterprise/list_domains',
-            rs_type=enterprise_pb2.ListDomainsResponse
-        )
-        return domain in rs.domain if rs.domain else False
-    
-    def _display_token_instructions(self, domain, token):
-        logging.info(f'\n{bcolors.OKGREEN}Token generated successfully!{bcolors.ENDC}\n')
-        logging.info(f'Domain: {bcolors.BOLD}{domain}{bcolors.ENDC}')
-        logging.info(f'Token:  {bcolors.BOLD}{token}{bcolors.ENDC}\n')
-        logging.info('Next steps:')
-        logging.info('1. Log into your domain registrar or DNS provider')
-        logging.info(f'2. Add a TXT record for domain "{domain}" with value:')
-        logging.info(f'   {bcolors.WARNING}{token}{bcolors.ENDC}')
-        logging.info('3. Wait for DNS propagation (may take a few minutes)')
-        logging.info(f'4. Run: domain reserve --action add --domain {domain}')
-    
-    def _refresh_enterprise_data(self, params, action_past_tense):
-        try:
-            api.query_enterprise(params)
-        except Exception as refresh_error:
-            logging.warning(f'Successfully {action_past_tense} domain but failed to refresh enterprise data: {refresh_error}')
-    
-    def _handle_api_error(self, error, domain, action, output_format):
-        error_code = error.result_code if hasattr(error, 'result_code') else 'Unknown'
-        
-        if DomainManagementHelper.is_feature_unavailable(error_code):
-            result = DomainManagementHelper.handle_unavailable_feature(output_format)
-            if result:
-                return result
-            return
-        
-        error_msg = DomainManagementHelper.get_error_message(error_code, domain, action)
-        
-        if output_format == 'json':
-            return json.dumps({
-                'message': error_msg,
-                'domain': domain,
-                'action': action,
-            }, indent=2)
-        
-        logging.error(error_msg)
